requirements.txt
Django==2.1.2
django-filter==2.0.0
djangorestframework==3.8.2
Markdown==3.0.1
gunicorn==19.9.0
psycopg2==2.7.5
dj-database-url==0.5.0
whitenoise==4.1
opencv-python==3.4.3.18
#dlib==19.16.0